limpeza completa e renderização condicional - done
